allocators: split allocators and pagelocker
>>> backports bitcoin@d7d187e8a451ae946fa14cead7962edbe0046f12

Pagelocker is only needed for secure (usually wallet) operations, so
don't make
the zero-after-free allocator depend on it.
