+++ src/key.h
@@ -7,8 +7,8 @@
 #ifndef PIVX_KEY_H
 #define PIVX_KEY_H
 
-#include "allocators.h"
 #include "serialize.h"
+#include "support/allocators/secure.h"
 #include "uint256.h"
 #include "pubkey.h"
 
@@ -20,7 +20,6 @@ class CPubKey;
 struct CExtPubKey;
 
 /**
- * secure_allocator is defined in allocators.h
  * CPrivKey is a serialized private key, with all parameters included
  * (PRIVATE_KEY_SIZE bytes)
  */
