+++ src/support/pagelocker.cpp
@@ -2,7 +2,11 @@
 // Distributed under the MIT/X11 software license, see the accompanying
 // file COPYING or http://www.opensource.org/licenses/mit-license.php.
 
-#include "allocators.h"
+#include "support/pagelocker.h"
+
+#if defined(HAVE_CONFIG_H)
+#include "config/pivx-config.h"
+#endif
 
 #ifdef WIN32
 #ifdef _WIN32_WINNT
