+++ src/utilstrencodings.h
@@ -10,7 +10,7 @@
 #ifndef BITCOIN_UTILSTRENCODINGS_H
 #define BITCOIN_UTILSTRENCODINGS_H
 
-#include "allocators.h"
+#include "support/allocators/secure.h"
 #include <stdint.h>
 #include <string>
 #include <vector>
